Split routes/activity.ts into feature-owned modules
The 779-line activity.ts held 14 handlers spanning three distinct
features. Move them into routes/activity/ mirroring the existing
routes/agents/ layout: metrics-routes.ts (heatmap, stats, daily
status, active hours, agents created, working time), token-routes.ts
(token stats/daily/by-project/by-model, harvest), history-routes.ts
(projects, agent list, agent detail), with shared.ts holding the deps
type and index.ts composing registration. Handler bodies are moved
verbatim; no behavior change.
